requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
apscheduler
requests
jinja2
gunicorn
psycopg2-binary